+++ src/components/RestaurantDetail/UserLike.js
@@ -1,41 +1,52 @@
 import { useEffect } from 'react'
 import { useContext, useState } from 'react'
-import { Button, Tooltip, OverlayTrigger } from 'react-bootstrap'
+import { Button, Modal } from 'react-bootstrap'
 import { Context } from '../../App'
+import OutreachButtons from '../OutreachButtons'
 import ProfileCard from '../ProfileCard'
- 
- 
 function UserLike({ user }) {
-  const { loggedInUser, dispatchUser } = useContext(Context)
-  const showInvite = false
-  const showMessage = false
+   const { loggedInUser, dispatchUser } = useContext(Context)
+   const showInvite = false
+   const showMessage = false
  
-  function friends() {
-     if(loggedInUser.response.friends.find(friend => friend._id === user._id)) return true
-     else return false
-  }
- 
-  function clickHandler(e) {
-     // show appropriate modal based on friend status
-     if(e.target !== window) friends() ? showMessage = true : showInvite = true
- 
-     // hide modal if user clicks outside of it
-     else {
-        showMessage = false
-        showInvite = false
-     }
-  }
- 
- 
- return (
+   const [show, setShow] = useState(false);
+   const handleClose = () => setShow(false);
+   const handleShow = () => setShow(true);
+   
+   function friends() {
+      if(loggedInUser.response.friends.find(friend => friend._id === user._id)) return true
+      else return false
+   }
+
+   function clickHandler(e) {
+
+      if(e.target !== window) friends() ? showMessage = true : showInvite = true
+      else {
+         showMessage = false
+         showInvite = false
+      }
+   }
+
+   return (
    <div style={{display:'flex', width:'auto', marginRight:"1%"}}>
       <div>
-         <ProfileCard username={user.username} profileImg={user.profileimg} user={user}/>
-         {/* Add modals which call SendFriendRequest or SendMessage here */}
+         <div onClick={handleShow}>
+            <ProfileCard username={user.username} profileImg={user.profileimg}/>
+         </div>
+         <Modal show={show} onHide={handleClose}>
+            <Modal.Header closeButton>
+               <Modal.Title>{user.username}</Modal.Title>
+            </Modal.Header>
+            <Modal.Body>{user.about}</Modal.Body>
+            <Modal.Footer style={{display:"flex", flexDirection:'column', justifyContent:"center", alignItems:"center"}}>
+               <OutreachButtons friends={friends()} user={user}/>
+            </Modal.Footer>
+         </Modal>
       </div>
    </div>
- )
+   )
 }
+
  
 export default UserLike
